[ADD] Movement class
Added the movement class for represent the movement system in the game. This class includes five method, all the getters and setters for the position and acceleration and the update method. It also include one operator =. The update method adjusts the position in function of the MoveDirection and the speed given by the Accelration class.